+++ AlchemistCode/Cards/Basic/Prime.cs
@@ -10,12 +10,11 @@ namespace Alchemist.AlchemistCode.Cards.Basic;
 
 public class Prime : AlchemistCard, ITranscendenceCard
 {
-    protected override bool IsGambitCard => true;
+    protected override ReactionCondition Reaction => ReactionCondition.Attack;
 
-    public Prime() : base(1, CardType.Attack, CardRarity.Basic, TargetType.AnyEnemy)
+    public Prime() : base(1, CardType.Skill, CardRarity.Basic, TargetType.Self)
     {
-        WithDamage(6, 2);
-        WithBlock(7, 3);
+        WithBlock(4, 3);
         WithTips(_ => Infusion.InfuseTips());
     }
 
@@ -25,9 +24,8 @@ public class Prime : AlchemistCard, ITranscendenceCard
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_attack_slash")).Execute(choiceContext);
-        await Infusion.InfuseChosen(choiceContext, this, PileType.Hand, 1);
-        if (IsReduced)
-            await CommonActions.CardBlock(this, play);
+        var reacted = ReactionActive;
+        await CommonActions.CardBlock(this, play);
+        await Infusion.InfuseChosen(choiceContext, this, PileType.Hand, reacted ? 2 : 1);
     }
 }