+++ AlchemistCode/Cards/Common/EchoStrike.cs
@@ -1,0 +1,22 @@
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+
+namespace Alchemist.AlchemistCode.Cards.Common;
+
+public class EchoStrike : AlchemistCard
+{
+    public EchoStrike() : base(1, CardType.Attack, CardRarity.Common, TargetType.AnyEnemy)
+    {
+        WithDamage(6, 2);
+    }
+
+    internal override bool GainsEffectWhenEnchanted => true;
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_attack_slash")).Execute(choiceContext);
+        if (IsEnchanted)
+            await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_attack_slash")).Execute(choiceContext);
+    }
+}